+++ lib/hi_energy_ai/resource.rb
@@ -33,5 +33,17 @@ module HiEnergyAi
     def app_post(path, params: {}, body: nil)
       client.app_post(path, params: params, body: body)
     end
+
+    # Used by mutating methods that accept either a positional Hash of
+    # body attributes (legacy) or keyword arguments (idiomatic Ruby).
+    # When a positional hash is given, kwargs are treated as query-string
+    # params; when only kwargs are given, they are all body attributes.
+    def split_attributes(attributes, params)
+      if attributes.nil?
+        [params, {}]
+      else
+        [attributes, params]
+      end
+    end
   end
 end