+++ spec/hi_energy_ai/client_spec.rb
@@ -114,5 +114,43 @@ RSpec.describe HiEnergyAi::Client do
 
       expect(dry_client.deals.list).to be_success
     end
+
+    it "accepts the preferred `server_dry_run:` alias" do
+      dry_client = described_class.new(api_key: api_key, base_url: base_url, server_dry_run: true)
+
+      stub_request(:get, "#{base_url}/deals")
+        .with(query: hash_including("dry_run" => "true"))
+        .to_return(status: 200, headers: { "Content-Type" => "application/json" }, body: "{}")
+
+      expect(dry_client.deals.list).to be_success
+      expect(dry_client.config.server_dry_run).to be(true)
+    end
+  end
+
+  describe "mutating resources" do
+    let(:body_matcher) { ->(req) { JSON.parse(req.body) == { "contact" => { "email" => "x@y.com" } } } }
+
+    it "accepts keyword-style attributes on create" do
+      stub_request(:post, "#{base_url}/contacts")
+        .with(&body_matcher)
+        .to_return(status: 201, headers: { "Content-Type" => "application/json" }, body: { data: { id: "1" } }.to_json)
+
+      expect(client.contacts.create(email: "x@y.com")).to be_success
+    end
+
+    it "still accepts a positional Hash on create (backwards compat)" do
+      stub_request(:post, "#{base_url}/contacts")
+        .with(&body_matcher)
+        .with(query: hash_including("foo" => "bar"))
+        .to_return(status: 201, headers: { "Content-Type" => "application/json" }, body: { data: { id: "1" } }.to_json)
+
+      expect(client.contacts.create({ email: "x@y.com" }, foo: "bar")).to be_success
+    end
+  end
+
+  describe "Tags resource" do
+    it "no longer exposes #search (use #list)" do
+      expect(client.tags).not_to respond_to(:search)
+    end
   end
 end